# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NetflixAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/TextItem": {
      "get": {
        "tags": [
          "TextItem"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextItem"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextItem"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TextItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TextItem"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/TextItem/{id}": {
      "get": {
        "tags": [
          "TextItem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TextItem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TextItem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/TextItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TextItem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/register": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TextItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "onPageTextId": {
            "type": "string",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "emailAddress": {
            "type": "string",
            "nullable": true
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "activated": {
            "type": "boolean",
            "nullable": true
          },
          "blockedUntil": {
            "type": "string",
            "nullable": true
          },
          "subscriptionId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "languageId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "profiles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "failedLoginAttempts": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "userHasInvited": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "userHasBeenInvited": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}